Move tests to PythonUDFSuite.
