+++ sql/core/src/test/scala/org/apache/spark/sql/DataFrameSuite.scala
@@ -2189,50 +2189,4 @@ class DataFrameSuite extends QueryTest with SharedSQLContext {
           |*(1) Range (0, 10, step=1, splits=2)""".stripMargin))
     }
   }
-
-  test("SPARK-28445: PythonUDF in grouping key and aggregate expressions") {
-    import IntegratedUDFTestUtils._
-
-    val scalaTestUDF = TestScalaUDF(name = "scalaUDF")
-    val pythonTestUDF = TestPythonUDF(name = "pyUDF")
-    assume(shouldTestPythonUDFs)
-
-    withTempView("testData") {
-      sql(
-        """CREATE OR REPLACE TEMPORARY VIEW testData AS
-          |SELECT * FROM VALUES
-          |(1, 1), (1, 2), (2, 1), (2, 2), (3, 1), (3, 2), (null, 1), (3, null), (null, null)
-          |AS testData(a, b)""".stripMargin)
-
-      val base = spark.table("testData")
-
-      val df = base.groupBy(scalaTestUDF(base("a") + 1))
-        .agg(scalaTestUDF(base("a") + 1), scalaTestUDF(count(base("b"))))
-      val df2 = base.groupBy(pythonTestUDF(base("a") + 1))
-        .agg(pythonTestUDF(base("a") + 1), pythonTestUDF(count(base("b"))))
-      checkAnswer(df, df2)
-
-      val df3 = base.groupBy(scalaTestUDF(base("a") + 1))
-        .agg(scalaTestUDF(base("a") + 1) + 1, scalaTestUDF(count(base("b"))))
-      val df4 = base.groupBy(pythonTestUDF(base("a") + 1))
-        .agg(pythonTestUDF(base("a") + 1) + 1, pythonTestUDF(count(base("b"))))
-      checkAnswer(df3, df4)
-
-      // PythonUDF in aggregate expression has grouping key in its arguments.
-      val df5 = base.groupBy(scalaTestUDF(base("a") + 1))
-        .agg(scalaTestUDF(scalaTestUDF(base("a") + 1)), scalaTestUDF(count(base("b"))))
-      val df6 = base.groupBy(pythonTestUDF(base("a") + 1))
-        .agg(pythonTestUDF(pythonTestUDF(base("a") + 1)), pythonTestUDF(count(base("b"))))
-      checkAnswer(df5, df6)
-
-      // PythonUDF over grouping key is argument to aggregate function.
-      val df7 = base.groupBy(scalaTestUDF(base("a") + 1))
-        .agg(scalaTestUDF(scalaTestUDF(base("a") + 1)),
-          scalaTestUDF(count(scalaTestUDF(base("a") + 1))))
-      val df8 = base.groupBy(pythonTestUDF(base("a") + 1))
-        .agg(pythonTestUDF(pythonTestUDF(base("a") + 1)),
-          pythonTestUDF(count(pythonTestUDF(base("a") + 1))))
-      checkAnswer(df7, df8)
-    }
-  }
 }
